DBZ-8582: Added test case for vector data type support in Maria DB.
